building the form module
