+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/BillOfLading/BillOfLadingTemplate.razor.cs
@@ -27,9 +27,9 @@
         MutableDataString = MutableDataString
       });
 
-      WebThreeState.CollectedFormValues.FormValues = response.CollectedFormValues.FormValues;
-      WebThreeState.CollectedFormValues.MutableDataString = response.CollectedFormValues.MutableDataString;
-
+      WebThreeState.ImmutableObject = response.ImmutableObject;
+      WebThreeState.MutableDataString = response.MutableDataString;
+     
      
     }
