+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/RequestForProposal/RequestForProposalTemplate.razor.cs
@@ -3,20 +3,27 @@
   using Microsoft.AspNetCore.Components;
   using nt.Client.Features.Base.Components;
   using nt.Client.Features.WebThree.Actions.AddFormDataToState;
-  using nt.Client.Features.WebThree.Components.NftTemplates;
   using System;
-  using System.Threading.Tasks;
 
   public class RequestForProposalTemplateModel : BaseComponent
   {
     [Parameter]
     public RequestForProposalData FormData { get; set; } = new RequestForProposalData();
-    
+
     public string MutableDataString { get; set; }
 
     public RequestForProposalTemplateModel()
     {
       FormData.Title = "Request For Proposal";
+    }
+
+    public void ConsoleData()
+    {
+      Console.WriteLine("Trying to Console...empty?");
+      foreach (System.Reflection.PropertyInfo prop in FormData.GetType().GetProperties())
+      {
+        Console.WriteLine($"{prop.Name}: {prop.GetValue(FormData)}");
+      }
     }
 
     public async void SendDataToState()
@@ -28,19 +35,8 @@
         MutableDataString = MutableDataString
       });
 
-      WebThreeState.CollectedFormValues.FormValues = response.CollectedFormValues.FormValues;
-      WebThreeState.CollectedFormValues.MutableDataString = response.CollectedFormValues.MutableDataString;
-
-     
-    }
-
-    public void ConsoleData()
-    {
-      Console.WriteLine("Trying to Console...empty?");
-      foreach (System.Reflection.PropertyInfo prop in FormData.GetType().GetProperties())
-      {
-        Console.WriteLine($"{prop.Name}: {prop.GetValue(FormData)}");
-      }
+      WebThreeState.ImmutableObject = response.ImmutableObject;
+      WebThreeState.MutableDataString = response.MutableDataString;
     }
   }
 }
