+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/RequestforQuotation/RequestForQuotationTemplate.razor.cs
@@ -4,18 +4,26 @@
   using nt.Client.Features.Base.Components;
   using nt.Client.Features.WebThree.Actions.AddFormDataToState;
   using System;
-  using System.Threading.Tasks;
 
   public class RequestForQuotationTemplateModel : BaseComponent
   {
     [Parameter]
     public RequestForQuotationData FormData { get; set; } = new RequestForQuotationData();
-    
+
     public string MutableDataString { get; set; }
 
     public RequestForQuotationTemplateModel()
     {
       FormData.Title = "Request For Quotation";
+    }
+
+    public void ConsoleData()
+    {
+      Console.WriteLine("Trying to Console...empty?");
+      foreach (System.Reflection.PropertyInfo prop in FormData.GetType().GetProperties())
+      {
+        Console.WriteLine($"{prop.Name}: {prop.GetValue(FormData)}");
+      }
     }
 
     public async void SendDataToState()
@@ -27,19 +35,8 @@
         MutableDataString = MutableDataString
       });
 
-      WebThreeState.CollectedFormValues.FormValues = response.CollectedFormValues.FormValues;
-      WebThreeState.CollectedFormValues.MutableDataString = response.CollectedFormValues.MutableDataString;
-
-     
-    }
-
-    public void ConsoleData()
-    {
-      Console.WriteLine("Trying to Console...empty?");
-      foreach (System.Reflection.PropertyInfo prop in FormData.GetType().GetProperties())
-      {
-        Console.WriteLine($"{prop.Name}: {prop.GetValue(FormData)}");
-      }
+      WebThreeState.ImmutableObject = response.ImmutableObject;
+      WebThreeState.MutableDataString = response.MutableDataString;
     }
   }
 }
