+++ Source/nt.Client/Pages/MintNftPage.razor.cs
@@ -2,10 +2,8 @@
 {
   using nt.Client.Features.Base.Components;
   using nt.Client.Features.WebThree;
-  using nt.Client.Features.WebThree.Actions.AddFormDataToState;
   using nt.Client.Features.WebThree.Components.NftTemplates;
   using nt.Shared.Features.WebThree;
-  using System.Threading.Tasks;
 
   public class MintNftPageModel : BaseComponent
   {
@@ -22,7 +20,8 @@
       set { }
     }
 
-    public FormDto FormData => WebThreeState.CollectedFormValues;
+    public ImmutableObjectBase FormData => WebThreeState.ImmutableObject;
 
+    public string MutableDataString => WebThreeState.MutableDataString;
   }
 }
